+++ VisionStepper.h
@@ -9,7 +9,7 @@ class VisionStepper {
     void init();
     void initPins(int enablePin, int directionPin, int stepPin);
     void initDelays(int startSpeedDelay, int highPhaseDelay, int maxSpeedDelay);
-    void initSizes(int wheelDiameter, int wheelRevolutionSteps);
+    void initSizes(float wheelDiameter, int wheelRevolutionSteps);
     void doLoop();
     void toggleDirection();
     void setDirectionForward();
